Refactor code and implement signing for macos
For signing on macos all the file changing (plist, package.json, etc) 
had to happen before signature otherwise the checksum wouldn't work. 
This was the drive for the refactoring. The signing part through the 
`desktop-packager` might disapear as this seems to be a very platform 
specific thing that is not worth to do right now before we have a full 
automated process to build and sign (it's easier to run manually the 
command line since this will have to be done manually anyway for now)
