+++ commands/darwin/package.js
@@ -1,32 +1,12 @@
 const path = require('path')
 const fs = require('fs').promises
-const execute = require('../utils/execute')
+const execute = require('../../utils/execute')
 const appdmg = require('appdmg')
 const plist = require('plist')
 
-const assetsFolder = path.resolve(__dirname, '..', 'assets', 'darwin')
+const assetsFolder = path.resolve(__dirname, '..', '..', 'assets', 'darwin')
 const dmgTemplatePath = path.resolve(assetsFolder, 'dmg.json.template')
 const dmgConfigPath = path.resolve(assetsFolder, 'dmg.json')
-
-/**
- * Register URL Scheme on OSX by updating `Info.plist` file.
- */
-const registerUrlScheme = async (src, appPkg) => {
-	// Register the app url scheme, by modifying the Info.plist
-	const plistPath = path.resolve(
-		src,
-		`${appPkg['executable-name']}.app`,
-		'Contents',
-		'Info.plist'
-	)
-	const plistFile = await fs.readFile(plistPath, 'utf8')
-	const plistObject = plist.parse(plistFile.toString())
-	plistObject.CFBundleURLTypes.push({
-		CFBundleURLName    : `${appPkg['executable-name']} URL`,
-		CFBundleURLSchemes : [appPkg['url-scheme']]
-	})
-	await fs.writeFile(plistPath, plist.build(plistObject))
-}
 
 const bakeDmgConfig = async (src, appPkg) => {
 	const templateFile = await fs.readFile(dmgTemplatePath)
@@ -61,7 +41,13 @@ const createDmg = async (outputInstallerPath) => {
 
 module.exports = async (src, appPkg, outputInstallerPath) => {
 	console.log('packaging for darwin')
-	await registerUrlScheme(src, appPkg)
 	await bakeDmgConfig(src, appPkg)
+	// Remove previous `.dmg` if it exists
+	try {
+		await fs.access(outputInstallerPath)
+		await fs.unlink(outputInstallerPath)
+	} catch(e) {
+		// File does not exist
+	}
 	await createDmg(outputInstallerPath)
 }
