+++ utils/modulePath.js
@@ -1,4 +1,5 @@
 const path = require('path')
 
-module.exports = module =>
+module.exports = (module) => {
 	path.resolve(require.resolve(path.join(module, 'package.json')), '..')
+}
